requirements.txt
absl-py==2.0.0
annotated-types==0.6.0
anyio==3.7.1
astunparse==1.6.3
async-timeout==4.0.3
beautifulsoup4==4.12.2
Brotli==1.1.0
cachetools==5.3.2
certifi==2023.11.17
charset-normalizer==3.3.2
click==7.1.2
colorama==0.4.6
dacite==1.8.1
Deprecated==1.2.14
exceptiongroup==1.2.0
fastapi==0.103.2
ffmpeg-python==0.2.0
flatbuffers==1.12
future==0.18.3
gast==0.4.0
google-auth==2.25.2
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.60.0
h11==0.12.0
h2==4.1.0
h5py==3.10.0
hpack==4.0.0
httpcore==0.13.7
httpx==0.19.0
hyperframe==6.0.1
idna==3.6
importlib-metadata==7.0.0
jaconv==0.3.4
keras==2.9.0
Keras-Preprocessing==1.1.2
libclang==16.0.6
Markdown==3.5.1
markdown-it-py==3.0.0
MarkupSafe==2.1.3
mdurl==0.1.2
mutagen==1.47.0
norbert==0.2.1
numpy==1.26.2
oauthlib==3.2.2
opt-einsum==3.3.0
packaging==23.2
pandas==1.5.3
platformdirs==3.11.0
protobuf==3.19.6
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycryptodomex==3.19.0
pydantic==2.5.2
pydantic_core==2.14.5
Pygments==2.17.2
pykakasi==2.2.1
python-dateutil==2.8.2
python-dotenv==1.0.0
python-slugify==8.0.1
pytube==15.0.0
pytz==2023.3.post1
rapidfuzz==3.5.2
redis==5.0.1
requests==2.31.0
requests-oauthlib==1.3.1
rfc3986==1.5.0
rich==13.7.0
rsa==4.9
scipy==1.11.4
six==1.16.0
sniffio==1.3.0
soundcloud-v2==1.3.1
soupsieve==2.5
spleeter==2.4.0
spotdl==4.2.4
spotipy==2.23.0
starlette==0.27.0
syncedlyrics==0.7.0
tensorboard==2.9.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.9.3
tensorflow-estimator==2.9.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.4.0
text-unidecode==1.3
typer==0.3.2
typing_extensions==4.9.0
tzdata==2023.3
Unidecode==1.3.7
urllib3==2.1.0
uvicorn==0.23.2
websockets==12.0
Werkzeug==3.0.1
wrapt==1.16.0
yt-dlp==2023.11.16
ytmusicapi==1.3.2
zipp==3.17.0
